fix script

+++ experiments/scan_features.py
@@ -24,10 +24,10 @@ class FeatureScanner:
     def __init__(self, sae, top_k=10):
         self.top_k = top_k
         
-        # Cache SAE params
-        self.W_enc = sae.W_enc.data.clone().detach()
-        self.b_enc = sae.b_enc.data.clone().detach()
-        self.b_dec = sae.b_dec.data.clone().detach()
+        # Cache SAE params (convert to float16 to match model)
+        self.W_enc = sae.W_enc.data.clone().detach().half()
+        self.b_enc = sae.b_enc.data.clone().detach().half()
+        self.b_dec = sae.b_dec.data.clone().detach().half()
         
         self.n_features = self.W_enc.shape[1]
